Implement file handling
Added RedCalls for file open and close. Added RedCalls for linking a
buffer with a file, reading from file into buffer, and writing back to
the file from the buffer. Added lua library calls and tables for use
within the editor.

Updated panic handling and fixed printing of panics to the terminal
after cleaning up the display logic.