Place buffers behind ContentBuffer trait
ContentBuffer trait supports all functions needed for a buffer to be
usable within the BadRed editor. EditorBuffer supports editor-specific
data tracking such as content and render dirty flags, but otherwise
serves as a passthrough to an underlying buffer that it holds
dynamically.

The first underlying buffer is a updated NaiveBuffer which holds its
content in a pure string and performs its insertions inefficiently
through string insertion.