Implement GapBuffer in bad_gap library
GapBuffer holds content with all pre-cursor elements placed at the front
of the vector and the post-cursor elements at the very back allowing for
a gap where the buffer is. This allows insertion to be O(1) at the
cursor. Moving the cursor itself is O(N).

GapBuffer has been implemented in its own bad_gap library for reuse and
better documentation.